@@ -1,4 +1,6 @@
-.fullDroptarget {
+import styled from 'styled-components'
+
+export const DroptargetWrapper = styled.div`
   position:absolute;
   top: 0;
   left: 0;
@@ -10,5 +12,4 @@
   display: flex;
   justify-content: center;
   align-items: center;
-
-}
+`
