@@ -1,19 +1,20 @@
 import React from 'react'
-import './sidebar.min.css'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import {
   faCheckCircle,
   faTimesCircle
 } from '@fortawesome/free-solid-svg-icons'
 
+import * as S from './styled'
+
 const Sidebar = props => {
   const { handleFullscreen, handleAutoplay, onPlay, onClear } = props
   return (
     <div className='item sidebar'>
       <div className='checkBoxWrapper'>
-        <ul className='video-options-btn'>
-          <li>
-            <input
+        <S.VideoOptions>
+          <S.VideoOptionsItem>
+            <S.VideoOptionsInput
               type='checkbox'
               id='checkboxOne'
               onChange={handleFullscreen}
@@ -21,10 +22,10 @@ const Sidebar = props => {
             />
             <FontAwesomeIcon style={props.videoOpts.fullscreen === 1 ? { opacity: '1' } : { opacity: '0' }} className='sidebar-fa-icon' width='1.325em' icon={faCheckCircle} />
             <FontAwesomeIcon style={props.videoOpts.fullscreen === 1 ? { opacity: '0' } : { opacity: '1' }} className='sidebar-fa-icon' width='1.325em' icon={faTimesCircle} />
-            <label htmlFor='checkboxOne'>Fullscreen</label>
-          </li>
-          <li>
-            <input
+            <S.VideoOptionsLabel htmlFor='checkboxOne'>Fullscreen</S.VideoOptionsLabel>
+          </S.VideoOptionsItem>
+          <S.VideoOptionsItem>
+            <S.VideoOptionsInput
               defaultChecked
               type='checkbox'
               id='checkboxTwo'
@@ -33,20 +34,20 @@ const Sidebar = props => {
             />
             <FontAwesomeIcon style={props.videoOpts.autoplay === 1 ? { opacity: '1' } : { opacity: '0' }} className='sidebar-fa-icon' width='1.325em' icon={faCheckCircle} />
             <FontAwesomeIcon style={props.videoOpts.autoplay === 1 ? { opacity: '0' } : { opacity: '1' }} className='sidebar-fa-icon' width='1.325em' icon={faTimesCircle} />
-            <label htmlFor='checkboxTwo'>Autoplay</label>
-          </li>
-        </ul>
+            <S.VideoOptionsLabel htmlFor='checkboxTwo'>Autoplay</S.VideoOptionsLabel>
+          </S.VideoOptionsItem>
+        </S.VideoOptions>
       </div>
 
-      <div className='btn-wrapper'>
-        <button className='fancy-button btn-play' onClick={onPlay}>
+      <S.BtnWrapper>
+        <S.Button className='fancy-button btn-play' onClick={onPlay}>
           <span className='ripple'>Play</span>
-        </button>
+        </S.Button>
 
-        <button className='fancy-button btn-clear ' onClick={onClear}>
+        <S.Button className='fancy-button btn-clear ' onClick={onClear}>
           <span className='ripple'>Clear</span>
-        </button>
-      </div>
+        </S.Button>
+      </S.BtnWrapper>
     </div>
   )
 }
